Sort order of offers appropriately by min/max range
This change fixes an issue with sorting the offer list when the amount
is shown as a range.  In OfferBookView::activate() we add a listener
for the sortTypeProperty on amountColumn and volumeColumn. When the
sortType is changed we set the comparator to be the approprate property
of the Offer; either getAmount/getMinAmount; getVolume/getMinVolume.

Fixes #3818