Replace the Get Support with Open Trader Chat until trade period is over
Showing an `Open Trader Chat` button until the trade period is over will
reduce the number of unnecessary support tickets and encourage
trader-to-trader dialogue.  If the trade timer expires without
completing the button changes (as before) to open a mediation ticket.

Implementation of this feature requires the button in TradeStepView to
notify its parent TradeSubView which in turn notify its parent
PendingTradesView that trader chat is to be opened.  Basically a
callback interface is passed two levels down the GUI hierarchy.

Fixes #3801