Add check for disputes if donation address is valid.
We check if the donation address is matching any of the past DAO params
as the dispute might be opened after the DAO param has been changed.
We show warnings for mediators and refund agents when a case gets added
which has an invalid address as well as at closing the case. We do not
permit closing such a case.
We also check if the reported address from the dispute opener matches
with our own address. If not we throw an error (TODO might not be the
best way to handle it... also not tested how it behaves).

For testing the check in the trade protocol needs to be temporary
deactivated to allow a trade to complete. Also the prevention of the
closing if the mediation case needs to be deactivated to test refund
agent cases.
It should be tested with default DAO param for donation address, with
one changed and with a newly changed (seller uses past param).
All those cases are valid. Only if seller uses a non DAO address the
protection gets activated.
The case where the seller manipulates the address in the dispute
message is not tested yet and will also requite a bit of hacking to get
that case simulated.