When accepting an offer do not round the BTC amount outside range
Fixes 3871.  The BTC amount of an offer was being adjusted up in
certain cases due to a rounding calculation from the fiat value
and price.  This would prevent some offers from being taken.

There's an existing check to ensure that the amount is not adjusted
below the minimum, here we add a check to ensure that the amount
is not adjusted higher than the maximum defined in the offer.