Editing an offer should not allow the BTC amount to be changed
When editing an offer to adjust the price the fiat amount is
recalculated to reflect the new price.  The BTC amount should
not change.  Due to a bug, sometimes the BTC amounts were
changed due to a recalculation from the fiat amount.
This PR Fixes 2798 by disabling a call to recalculate BTC and
min BTC amounts when editing.