Limit MaxRAM for all java apps
Setting DEFAULT_JVM_OPTS="-XX:MaxRAM=4g" in bisq-desktop has not
resulted any reported issues, so this jvm option can be used in other
launch scripts to reduce the amount of memory reserved by headless
apps.  This change is based on the assumption none of the headless
apps require as much memory as desktop.