Fix QR code display of offer funding amount
When creating an offer the small QR code displayed for funding
was off by an amount that represent the mining fees.  This was due
to the QR code being displayed simultaneously with the wallet
calulating the fee.

We fix the issue by listening to the data model `missingCoin`
value of the offer object, and updating the QR code image
appropriately.

Fixes #3666
Fixes #4194