Use full sync between seednodes
If we use the diff sync between seed nodes we create a race condition
where when a seednode gets updated to the new system, it does not
sync up properly with the other seed nodes. And that would be fatal.

So for the time being, when a seednode asks for data, it uses the
"old" big requests with all object keys. Should not be a problem
for now since they have enough bandwidth.