Add pruning support for trade statistics
We add an interface for prunable PersistableNetworkPayloads so we can
handle it generically in the p2p module. We check at startup if our
persisted data has old elements and remove those. Data we receive from
the seed node at startup will get checked as well for old objects as
well data we receive from the network.